updated tests + RIP TestyMcTestFace
